[QMS-903] Extract shared itemNameColor() helper for tree delegates
CWksItemDelegate::paintProject/paintDevice/paintItem and
CMapItemDelegate::paint each derived the name/status text color from
the same colorRole/colorGroup/colorName logic. Factor it into
CDraw::itemNameColor(opt, isVisible) and update CLAUDE.md's delegate
cleanup checklist accordingly.
